+++ crates/rust-client/src/account/mod.rs
@@ -97,6 +97,7 @@ pub mod component {
         StorageValueName,
     };
     pub use miden_protocol::account::{AccountComponent, AccountComponentMetadata};
+    pub use miden_standards::account::access::Ownable2Step;
     pub use miden_standards::account::auth::*;
     pub use miden_standards::account::components::{
         basic_fungible_faucet_library,
@@ -107,7 +108,6 @@ pub mod component {
         singlesig_acl_library,
         singlesig_library,
     };
-    pub use miden_standards::account::access::Ownable2Step;
     pub use miden_standards::account::faucets::{
         BasicFungibleFaucet,
         NetworkFungibleFaucet,